% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullScenario.R
\name{fullScenario}
\alias{fullScenario}
\title{Creates Full Scenario based on a datasheet}
\usage{
fullScenario(ssimProject, scenarioName, tag_list)
}
\description{
This function will create a unique sub scenario based on a given datasheet.
}
