% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDatasheet.R
\name{loadDatasheet}
\alias{loadDatasheet}
\title{Allows to edit a datasheet from a given library}
\usage{
loadDatasheet(
  datasheetName,
  ssimObject,
  datasheetFolder = getwd(),
  params = NULL
)
}
\description{
This function wraps saveDatasheet() and datasheet()
}
