% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSheet.R
\name{loadSheet}
\alias{loadSheet}
\title{Allows to edit a datasheet from a given library}
\usage{
loadSheet(sheetname, proj_or_sce, path = getwd(), params = NULL)
}
\description{
This function wraps saveDatasheet() and datasheet()
}
