% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subScenario.R
\name{subScenario}
\alias{subScenario}
\title{Creates Sub-Scenario based on a datasheet}
\usage{
subScenario(
  ssimProject,
  tag = "default",
  datasheetName,
  datasheetParameters,
  savesheet = F,
  export = T,
  datasheetFolder
)
}
\description{
This function will create a unique sub scenario based on a given datasheet.
}
