% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editDatasheet.R
\name{editDatasheet}
\alias{editDatasheet}
\title{Allows to edit a datasheet from a given library}
\usage{
editDatasheet(
  datasheetName,
  ssimObject,
  argumentList,
  saveSheet = T,
  erase = F,
  save_as_CSV = F,
  CSV_folder = NULL
)
}
\description{
This function wraps saveDatasheet() and datasheet()
}
