% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editDatasheet.R
\name{editDatasheet}
\alias{editDatasheet}
\title{Allows to edit a datasheet from a given library}
\usage{
editDatasheet(
  datasheetName,
  tag = NULL,
  ssimObject,
  argumentList = NULL,
  saveSheet = T,
  erase = F,
  export = F,
  datasheetFolder = NULL
)
}
\description{
This function wraps saveDatasheet() and datasheet()
}
