% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullScenario.R
\name{fullScenario}
\alias{fullScenario}
\title{Create Full Scenario based on a set of Sub-Scenario with one or more tag(s)}
\usage{
fullScenario(ssimProject, scenarioName, tag)
}
\arguments{
\item{ssimProject}{(ssimObject) Syncrosim Project to save to.}

\item{scenarioName}{(character) The name of the full scenario.}

\item{tag}{(character or character vector) The tag(s) associated with the subscenarios, can be NULL, 
default to "default".}
}
\description{
Wraps around rsncrosim::scenario in order to Assemble a full scenario from 
a set of Sub-Scenario with one or more tag(s).
}
