% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDatasheet.R
\name{loadDatasheet}
\alias{loadDatasheet}
\title{Load a datasheet into library from CSV}
\usage{
loadDatasheet(
  datasheetName,
  tag = "default",
  ssimObject = NULL,
  datasheetFolder = NULL
)
}
\arguments{
\item{datasheetName}{(character) The name of the datasheet (ex: RunControl).}

\item{tag}{(character) The tag associated with that datasheet, can be NULL, default to "default".}

\item{ssimObject}{(ssimObject) Object of the class ssimObject from which to extract and}

\item{datasheetFolder}{(character) The path to where to save the datasheet if export is TRUE, default to NULL.}
}
\description{
Wraps rsyncrosim::saveDatasheet() to read in a load into library a datasheet directly
from a csv file.
}
